add the docker for build the book (#6)

+++ docs/book/install-and-build.sh
@@ -60,11 +60,13 @@ esac
 
 # grab mdbook
 # we hardcode linux/amd64 since rust uses a different naming scheme and it's a pain to tran
-echo "downloading mdBook-v0.3.1-${arch}-${target}.${ext}"
-set -x
-curl -sL -o /tmp/mdbook.${ext} https://github.com/rust-lang-nursery/mdBook/releases/download/v0.3.1/mdBook-v0.3.1-${arch}-${target}.${ext}
-${cmd} /tmp/mdbook.${ext}
-chmod +x /tmp/mdbook
+if [ ! -f /tmp/mdbook ]; then
+    echo "downloading mdBook-v0.3.1-${arch}-${target}.${ext}"
+    set -x
+    curl -sL -o /tmp/mdbook.${ext} https://github.com/rust-lang-nursery/mdBook/releases/download/v0.3.1/mdBook-v0.3.1-${arch}-${target}.${ext}
+    ${cmd} /tmp/mdbook.${ext}
+    chmod +x /tmp/mdbook
+fi
 
 echo "grabbing the latest released controller-gen"
 go get sigs.k8s.io/controller-tools/cmd/controller-gen@v0.3.0
